+++ src/ralph/backends/loader.py
@@ -1,8 +1,14 @@
 """Ralph backend loader."""
 
 import logging
+import sys
 from functools import lru_cache
-from typing import Dict, List, Tuple
+from typing import Dict, Tuple
+
+if sys.version_info < (3, 10):
+    from importlib_metadata import EntryPoints, entry_points
+else:
+    from importlib.metadata import EntryPoints, entry_points
 
 from ralph.backends.data.base import (
     AsyncListable,
@@ -13,70 +19,50 @@ from ralph.backends.data.base import (
     Writable,
 )
 from ralph.backends.lrs.base import BaseAsyncLRSBackend, BaseLRSBackend
-from ralph.utils import import_string
 
 logger = logging.getLogger(__name__)
 
-DATA_BACKENDS: List[str] = [
-    "ralph.backends.data.async_es.AsyncESDataBackend",
-    "ralph.backends.data.async_lrs.AsyncLRSDataBackend",
-    "ralph.backends.data.async_mongo.AsyncMongoDataBackend",
-    "ralph.backends.data.async_ws.AsyncWSDataBackend",
-    "ralph.backends.data.clickhouse.ClickHouseDataBackend",
-    "ralph.backends.data.es.ESDataBackend",
-    "ralph.backends.data.fs.FSDataBackend",
-    "ralph.backends.data.ldp.LDPDataBackend",
-    "ralph.backends.data.lrs.LRSDataBackend",
-    "ralph.backends.data.mongo.MongoDataBackend",
-    "ralph.backends.data.s3.S3DataBackend",
-    "ralph.backends.data.swift.SwiftDataBackend",
-]
-LRS_BACKENDS: List[str] = [
-    "ralph.backends.lrs.async_es.AsyncESLRSBackend",
-    "ralph.backends.lrs.async_mongo.AsyncMongoLRSBackend",
-    "ralph.backends.lrs.clickhouse.ClickHouseLRSBackend",
-    "ralph.backends.lrs.es.ESLRSBackend",
-    "ralph.backends.lrs.fs.FSLRSBackend",
-    "ralph.backends.lrs.mongo.MongoLRSBackend",
-]
-
 
 def get_backends(
-    backends: List[str], base_backends: Tuple[type, ...]
+    backends: EntryPoints, base_backends: Tuple[type, ...]
 ) -> Dict[str, type]:
     """Return sub-classes of `base_backends` from `backends`.
 
     Args:
-        backends (list of str): A list of dotted backend import paths.
-            Ex.: ["foo.bar.BackendA", "foo.bar.BackendB"].
+        backends (EntryPoints): EntryPoints pointing to backend classes.
+            Ex.: [EntryPoint(name="backend_name", value="foo.bar:BackendA"].
         base_backends (tuple of type): A tuple of base backend classes to filter for.
             Ex.: ("BaseDataBackend",)
 
     Return:
-        dict: A dictionary of backend classes by their name property.
+        dict: A dictionary of backend classes by their `EntryPoint.name` property.
             Ex.: {"fs": FSDataBackend}
     """
-    backend_classes = []
-    for backend in backends:
+    backend_classes = {}
+    for backend in sorted(backends, key=lambda x: x.name, reverse=True):
         try:
-            backend_class = import_string(backend)
+            backend_class = backend.load()
         except Exception as error:  # noqa: BLE001
-            logger.debug("Failed to import '%s' backend: %s", backend, error)
+            logger.debug(
+                "Failed to import '%s' backend from '%s': %s",
+                backend.name,
+                backend.value,
+                error,
+            )
             continue
 
         if issubclass(backend_class, base_backends):
-            backend_classes.append(backend_class)
-    return {
-        backend.name: backend
-        for backend in sorted(backend_classes, key=lambda x: x.name, reverse=True)
-    }
+            backend_classes[backend.name] = backend_class
+
+    return backend_classes
 
 
 @lru_cache(maxsize=1)
 def get_cli_backends() -> Dict[str, type]:
     """Return Ralph's backend classes for cli usage."""
     base_backends = (BaseAsyncDataBackend, BaseDataBackend)
-    return get_backends(DATA_BACKENDS, base_backends)
+    data_backends = entry_points(group="ralph.backends.data")
+    return get_backends(data_backends, base_backends)
 
 
 @lru_cache(maxsize=1)
@@ -102,4 +88,5 @@ def get_cli_list_backends() -> Dict[str, type]:
 @lru_cache(maxsize=1)
 def get_lrs_backends() -> Dict[str, type]:
     """Return Ralph's backend classes for LRS usage."""
-    return get_backends(LRS_BACKENDS, (BaseAsyncLRSBackend, BaseLRSBackend))
+    lrs_backends = entry_points(group="ralph.backends.lrs")
+    return get_backends(lrs_backends, (BaseAsyncLRSBackend, BaseLRSBackend))